ODA-419
Extended interface for (temporarely) handling localization (Cohn) combined with assimilating the observations one at a time. 
NOTE: the weights in the observer SHOULD corresponds to those in the model.
This is directly why this setup is far from ideal, will probably be removed in the future when we reimplement localization in OpenDA!

git-svn-id: https://svn.oss.deltares.nl/repos/openda/trunk@341 e6e3f6e9-ec65-4820-a5bc-431d91f1ef11

Former-commit-id: 641a663d08d345ea3385cdca71f4cf3dc514adb5